+++ packages/flatomise/src/main.ts
@@ -1,10 +1,11 @@
 /**
  * Flat promise that can be resolved or rejected from outside.
  */
-export interface Flatomise {
-  promise: Promise<void>;
-  resolve: () => void;
-  reject: () => void;
+export interface Flatomise<T> {
+  promise: Promise<T>;
+  resolve: (value: T | PromiseLike<T>) => void;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  reject: (reason?: any) => void;
 }
 
 /**
@@ -21,11 +22,11 @@ export interface Flatomise {
  * flatomise.resolve();
  * ```
  */
-export function newFlatomise(): Flatomise {
-  const flatomise: Partial<Flatomise> = {};
-  flatomise.promise = new Promise<void>((resolve, reject) => {
+export function newFlatomise<T>(): Flatomise<T> {
+  const flatomise: Partial<Flatomise<T>> = {};
+  flatomise.promise = new Promise<T>((resolve, reject) => {
     flatomise.resolve = resolve;
     flatomise.reject = reject;
   });
-  return flatomise as Flatomise;
+  return flatomise as Flatomise<T>;
 }