+++ packages/logger/src/type.ts
@@ -1,4 +1,14 @@
+/**
+ * Represents the AlwatrLogger interface.
+ * The AlwatrLogger provides methods for logging various types of information.
+ */
 export interface AlwatrLogger {
+
+  /**
+   * Debug state for current scope base on localStorage `ALWATR_LOG` pattern.
+   */
+  debugMode: boolean;
+
   /**
    * `console.debug` property change.
    *