+++ packages/node-fs/src/make-file.ts
@@ -1,0 +1,18 @@
+import {open} from 'node:fs/promises';
+
+import {logger} from './common';
+
+/**
+ * Make empty file.
+ *
+ * @param path - file path
+ *
+ * @example
+ * ```ts
+ * await makeFile('./file.txt');
+ * ```
+ */
+export async function makeEmptyFile(path: string): Promise<void> {
+  logger.logMethodArgs?.('makeEmptyFile', '...' + path.slice(-32));
+  return (await open(path, 'w')).close();
+}