+++ packages/exit-hook/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/Alwatr/nanolib/compare/@alwatr/exit-hook@1.0.2...@alwatr/exit-hook@1.0.3) (2024-01-08)
+
+**Note:** Version bump only for package @alwatr/exit-hook
+
 ## [1.0.2](https://github.com/Alwatr/nanolib/compare/@alwatr/exit-hook@1.0.1...@alwatr/exit-hook@1.0.2) (2024-01-03)
 
 **Note:** Version bump only for package @alwatr/exit-hook