+++ packages/platform-info/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/Alwatr/nanolib/compare/@alwatr/platform-info@1.1.0...@alwatr/platform-info@1.1.1) (2024-01-08)
+
+**Note:** Version bump only for package @alwatr/platform-info
+
 # [1.1.0](https://github.com/Alwatr/nanolib/compare/@alwatr/platform-info@1.0.7...@alwatr/platform-info@1.1.0) (2024-01-03)
 
 ### Features